// src/app/api/models/cancellation-summary.ts
/* tslint:disable */
export interface CancellationSummary {
  allCardCount?: number;
  allCardTotal?: number;
  allRefundAmount?: number;
  collectionCardCount?: number;
  collectionCardTotal?: number;
  collectionRefundAmount?: number;
  date?: string;
  deliveryCardCount?: number;
  deliveryCardTotal?: number;
  deliveryRefundAmount?: number;
  storeId?: string;
}